+++ lib/models/package.rb
@@ -14,13 +14,25 @@ module OpenBuildServiceAPI
     end
 
     def title
+      return @title_updated if @title_updated
+
       title = meta.xpath('//package/title')
       title.empty? ? nil : title[0].text
     end
 
+    def title=(value)
+      @title_updated = value
+    end
+
     def description
+      return @description_updated if @description_updated
+
       description = meta.xpath('//package/description')
       description.empty? ? nil : description[0].text
+    end
+
+    def description=(value)
+      @description_updated = value
     end
 
     def meta(opts = {})
@@ -47,6 +59,25 @@ module OpenBuildServiceAPI
 
     def rebuild_failed!
       @connection.send_request(:post, "/build/#{CGI.escape(@project.name)}", cmd: :rebuild, package: @name, code: :failed)
+      true
+    end
+
+    def save!
+      @meta_reload = true
+      meta_xml = meta
+
+      meta_xml.xpath('//package/title')[0].content = @title_updated if @title_updated
+      meta_xml.xpath('//package/description')[0].content = @description_updated if @description_updated
+
+      @connection.send_request(:put, "/source/#{CGI.escape(@project.name)}/#{CGI.escape(@name)}/_meta", request_body: meta_xml.to_xml)
+
+      # reset updated values
+      @title_updated = nil
+      @description_updated = nil
+
+      # updated cached meta
+      @cached_meta = meta_xml.to_xml
+
       true
     end